+++ xpasses/XR005/testdata/src/a/schema/schema.go
@@ -1,0 +1,12 @@
+package schema
+
+type ReadFunc func(*ResourceData, interface{}) error
+
+type Resource struct {
+	Read   ReadFunc
+	Schema map[string]*Schema
+}
+
+type ResourceData struct{}
+
+type Schema struct{}